agent/requirements.txt
python-dotenv>=0.19.0
openai>=1.0.0
pandas_ta>=0.3.14b